chore: add `tx-mode` input for migration/schema apply (#329)
This pull request includes changes to add a new `tx-mode` input
parameter for specifying the transaction mode for migrations. The
changes are made across multiple files to ensure consistency and proper
documentation.

### Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R263):
Added the `tx-mode` parameter description under the sections for inputs
in the migration configuration.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R263)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R368)

### Configuration updates:

*
[`migrate/apply/action.yml`](diffhunk://#diff-f0659287a2a21eb4b307db47bfa1e9f3f93cbdbb1dcd4c57087940843d14d861R44-R47):
Introduced the `tx-mode` parameter with a description and default value.
*
[`schema/apply/action.yml`](diffhunk://#diff-77f4937173e7eea7a278bfea3273901d38d1b44ff7670486c1a59c31384b84bfR52-R55):
Added the `tx-mode` parameter with detailed description and default
value.