atlasaction: add a common function to handler approval flow (#340)
This pull request includes changes to the `atlasaction/action.go` file
to refactor the `MigrateDown` function and add a new helper function to
handle approval waiting logic. The most important changes include
simplifying the retry logic within `MigrateDown` and introducing the
`waitingForApproval` function which will be used by `schema/apply`

Refactoring and simplification:

*
[`atlasaction/action.go`](diffhunk://#diff-27a1debcaa5772fc7928822261793bc1bd5ebeca936105e889fbc6a9efc93ac5L386-R407):
Refactored the `MigrateDown` function to remove inline retry logic and
replace it with a call to the new `waitingForApproval` function. This
change simplifies the `MigrateDown` function and makes the code more
modular.
*
[`atlasaction/action.go`](diffhunk://#diff-27a1debcaa5772fc7928822261793bc1bd5ebeca936105e889fbc6a9efc93ac5R888-R916):
Added a new helper function `waitingForApproval` to encapsulate the
retry logic for waiting for plan approval. This function handles retry
intervals, timeouts, and stopping conditions.